Scores version 2.0
Core class revised.

Score_CLI.py added - This runs the program in the command line.
This takes the place of Zach_Scores.py which was previously a standalone.
This was done so that the base class could be used in the TK interface.

Score_Gui.py added - This is a tk interface for the program.
This should be the prefered way to run this program.